refactor: remove database folder and rename config.js to mysql.js
